+++ private/errorConsole.js
@@ -1,0 +1,17 @@
+'use strict';
+
+const { Console } = require('console');
+
+/**
+ * The `console` API, but all output is to `stderr`. This allows `console.group`
+ * to be used with `console.error`.
+ * @kind constant
+ * @name errorConsole
+ * @ignore
+ */
+const errorConsole = new Console({
+  stdout: process.stderr,
+  stderr: process.stderr,
+});
+
+module.exports = errorConsole;